+++ gateway/src/main/java/ru/practicum/shareit/booking/BookingController.java
@@ -48,7 +48,7 @@ public class BookingController {
 
 	@GetMapping("/{bookingId}")
 	public ResponseEntity<Object> getBooking(@RequestHeader("X-Sharer-User-Id") int userId,
-			@PathVariable Long bookingId) {
+			@PathVariable Integer bookingId) {
 		log.info("Get booking {}, userId={}", bookingId, userId);
 		return bookingClient.getBooking(userId, bookingId);
 	}
